Move editor toolbar to editor component
Editor toolbar that consists language drop down menu and code submit button moved to from collaboration component to editor component. A prop `isMainEditor` has been added to the editor component. The new prop is used to render the toolbar as we would only change the language or submit the code of the main editor (the other editor is view only from peer's code).